🚧 (flash): Begin to create the layout of component
