+++ demos/playgrounds/flash.html
@@ -2,7 +2,7 @@
 <div class="playground">
   <playground-element element="kami-flash">
     <template data-element="kami-flash" data-target="host">
-      <kami-flash></kami-flash>
+      <kami-flash message="Hello world"></kami-flash>
     </template>
   </playground-element>
 </div>
